fixing hybrid script issue. updating reranker server to run using gpu. xtra logging
