+++ api/run_configs/default.json
@@ -12,9 +12,10 @@
       "n_retrieval": 256,
       "n_keyword_srch": 20
     },
-    "reranker_service": "crossencoder",
+    "reranker_service": "remote",
     "reranker_args": {
-      "model_name_or_path": "mixedbread-ai/mxbai-rerank-large-v1"
+      "model_name_or_path": "mixedbread-ai/mxbai-rerank-large-v1",
+      "batch_size": 64
     },
     "paper_finder_args": {
       "n_rerank": 50,
