+++ api/scholarqa/rag/reranker/remote_reranker.py
@@ -14,13 +14,17 @@ class RemoteRerankerClient(AbstractReranker):
     
     def __init__(self, model_name_or_path: str = "mixedbread-ai/mxbai-rerank-large-v1", 
                  reranker_type: str = "crossencoder", 
+                 batch_size: int = 64,
                  timeout: float = 60.0):
+        # Get service URL from environment variable (set by Docker Compose)
         self.service_url = os.getenv("RERANKER_SERVICE_URL", "http://localhost:8001").rstrip('/')
         self.model_name_or_path = model_name_or_path
         self.reranker_type = reranker_type
+        self.batch_size = batch_size
         self.timeout = timeout
+        self.device = "remote"  # Indicate this is a remote service
         
-        logger.info(f"Initialized RemoteRerankerClient: {self.service_url}")
+        logger.info(f"Initialized RemoteRerankerClient: {self.service_url} with model: {model_name_or_path}, batch_size: {batch_size}")
         self._test_connection()
     
     def _test_connection(self):
@@ -43,7 +47,8 @@ class RemoteRerankerClient(AbstractReranker):
                 "query": query,
                 "passages": passages,
                 "model_name_or_path": self.model_name_or_path,
-                "reranker_type": self.reranker_type
+                "reranker_type": self.reranker_type,
+                "batch_size": self.batch_size
             }
             
             with httpx.Client(timeout=self.timeout) as client:
@@ -53,6 +58,9 @@ class RemoteRerankerClient(AbstractReranker):
                     raise Exception(f"Service error: {response.status_code} - {response.text}")
                 
                 result = response.json()
+                # Log the actual device being used by the remote service
+                if "device" in result:
+                    logger.info(f"Remote reranker using device: {result['device']}")
                 return result["scores"]
                 
         except httpx.TimeoutException:
