+++ docker-compose.yaml
@@ -18,7 +18,7 @@ services:
       - LITELLM_LOG=WARNING
       - CONFIG_PATH=run_configs/default.json
       # url of remote reranker service to be reached by the API container
-      - RERANKER_SERVICE_URL=http://host.docker.internal:8001
+      - RERANKER_SERVICE_URL=http://host.docker.internal:${RERANKER_PORT}
     ports:
       - 8000:8000
     env_file:
